Funcionalidades del sistema. ABM Rol.

+++ gd_clinica_medica_frba/objetos_de_bd_Cami.sql
@@ -3,19 +3,10 @@
 **************************************************************************************/
 
 IF OBJECT_ID ('LOS_TRIGGERS.Calendario') IS NOT NULL DROP TABLE LOS_TRIGGERS.Calendario
-
-SET ANSI_NULLS ON
-GO
-SET QUOTED_IDENTIFIER ON
-GO
 CREATE TABLE [LOS_TRIGGERS].[Calendario](
 	[dia_del_año] [date] NOT NULL,
- CONSTRAINT [PK_Calendar_Date] PRIMARY KEY CLUSTERED 
-(
-	[dia_del_año] ASC
-)WITH (PAD_INDEX = OFF, STATISTICS_NORECOMPUTE = OFF, IGNORE_DUP_KEY = OFF, ALLOW_ROW_LOCKS = ON, ALLOW_PAGE_LOCKS = ON) ON [PRIMARY]
-) ON [PRIMARY]
-
+ CONSTRAINT [PK_Calendar_Date] PRIMARY KEY CLUSTERED ([dia_del_año] ASC)
+ WITH (PAD_INDEX = OFF, STATISTICS_NORECOMPUTE = OFF, IGNORE_DUP_KEY = OFF, ALLOW_ROW_LOCKS = ON, ALLOW_PAGE_LOCKS = ON) ON [PRIMARY]) ON [PRIMARY]
 GO
 
 DECLARE @primer_dia datetime
@@ -24,29 +15,79 @@ SET @primer_dia = DATEADD(YEAR, DATEDIFF(YEAR, 0, GETDATE()), 0)
 SET @ultimo_dia = DATEADD(d, 365, @primer_dia)
 
 WHILE @primer_dia <= @ultimo_dia
-      BEGIN
-             INSERT INTO [LOS_TRIGGERS].[Calendario] (dia_del_año)
-				 SELECT @primer_dia SET @primer_dia = DATEADD(dd, 1, @primer_dia)
-      END
+	BEGIN
+		insert into [LOS_TRIGGERS].[Calendario] (dia_del_año)
+			select @primer_dia SET @primer_dia = DATEADD(dd, 1, @primer_dia)
+	END
 
 --- << ABM de Rol >> ---
-/*
-IF OBJECT_ID ('LOS_TRIGGERS.AltaRol') is not null DROP PROCEDURE LOS_TRIGGERS.AltaRol
+
+IF OBJECT_ID ('LOS_TRIGGERS.ModificarRol') is not null DROP PROCEDURE LOS_TRIGGERS.ModificarRol
 GO
-CREATE PROC LOS_TRIGGERS.AltaRol (@nombre_rol varchar(255)) AS
+CREATE PROC LOS_TRIGGERS.ModificarRol (@rol varchar(255), @nuevo_nombre varchar(255)) AS
 	BEGIN
+		--select TABLE_NAME From INFORMATION_SCHEMA.COLUMNS Where column_name = 'nombre_rol'
+		update LOS_TRIGGERS.Administrador set nombre_rol = @nuevo_nombre where nombre_rol=@rol
+		update LOS_TRIGGERS.Afiliado set nombre_rol = @nuevo_nombre where nombre_rol=@rol
+		update LOS_TRIGGERS.Profesional set nombre_rol = @nuevo_nombre where nombre_rol=@rol
+	END;
+GO
+
+
+IF OBJECT_ID ('LOS_TRIGGERS.AltaRolAdministrador') is not null DROP PROCEDURE LOS_TRIGGERS.AltaRolAdministrador
+GO
+CREATE PROC LOS_TRIGGERS.AltaRolAdministrador AS
+	BEGIN
+		IF OBJECT_ID ('LOS_TRIGGERS.Administrador') IS NOT NULL DROP TABLE LOS_TRIGGERS.Administrador
+		CREATE TABLE LOS_TRIGGERS.Administrador (
+			[admi_id] [numeric](18, 0) IDENTITY(1,1) NOT NULL primary key,
+			[admi_habilitacion] [bit] NULL
+		);
 	
 	END;
 GO
 
-IF OBJECT_ID ('LOS_TRIGGERS.ModificarRol') is not null DROP PROCEDURE LOS_TRIGGERS.ModificarRol
+IF OBJECT_ID ('LOS_TRIGGERS.AltaRolAfiliado') is not null DROP PROCEDURE LOS_TRIGGERS.AltaRolAfiliado
 GO
-CREATE PROC LOS_TRIGGERS.ModificarRol (@nombre varchar(255)) AS
+CREATE PROC LOS_TRIGGERS.AltaRolAfiliado AS
 	BEGIN
-
+	IF OBJECT_ID ('LOS_TRIGGERS.Afiliado') IS NOT NULL DROP TABLE LOS_TRIGGERS.Afiliado
+	CREATE TABLE LOS_TRIGGERS.Afiliado (
+		[afil_numero] [numeric](18, 0) IDENTITY(1,1) NOT NULL primary key,
+		[afil_estado_civil] [varchar](255) NULL,
+		[afil_habilitacion] [bit] NULL,
+		[afil_titular_grupo_familiar] [numeric](18, 0) NULL foreign key references LOS_TRIGGERS.Afiliado,
+		[afil_plan_medico] [numeric](18, 0) NULL foreign key references LOS_TRIGGERS.Plan_Medico,
+		[afil_cant_consultas_realizadas] [numeric](18, 0) NULL
+	);
+	
 	END;
 GO
-*/
+
+IF OBJECT_ID ('LOS_TRIGGERS.AltaRolProfesional') is not null DROP PROCEDURE LOS_TRIGGERS.AltaRolProfesional
+GO
+CREATE PROC LOS_TRIGGERS.AltaRolProfesional AS
+	BEGIN
+	IF OBJECT_ID ('LOS_TRIGGERS.Profesional ') IS NOT NULL DROP TABLE LOS_TRIGGERS.Profesional 
+	CREATE TABLE LOS_TRIGGERS.Profesional  (
+		[prof_id] [numeric](18, 0) IDENTITY(1,1) NOT NULL primary key,
+		[prof_matricula] [numeric](18, 0) NULL,
+		[prof_horas_laborales] [numeric](2, 0) NULL,
+		[prof_habilitacion] [bit] NULL,
+	);
+	
+	END;
+GO
+
+IF OBJECT_ID ('LOS_TRIGGERS.AltaRol') is not null DROP PROCEDURE LOS_TRIGGERS.AltaRol
+GO
+CREATE PROC LOS_TRIGGERS.AltaRol (@nombre_rol varchar(255)) AS
+	BEGIN
+		IF (@nombre_rol='Afiliado') EXEC LOS_TRIGGERS.AltaRolAfiliado
+		ELSE IF (@nombre_rol='Administrador') EXEC LOS_TRIGGERS.AltaRolAdministrador
+		ELSE IF (@nombre_rol='Profesional') EXEC LOS_TRIGGERS.AltaRolProfesional
+	END;
+GO
 
 IF OBJECT_ID ('LOS_TRIGGERS.InhabilitarRol') is not null DROP PROCEDURE LOS_TRIGGERS.InhabilitarRol
 GO
@@ -54,18 +95,18 @@ CREATE PROC LOS_TRIGGERS.InhabilitarRol (@nombre_rol varchar(255)) AS
 	BEGIN
 		IF (@nombre_rol='Afiliado')
 			BEGIN
-				UPDATE LOS_TRIGGERS.Afiliado set afil_habilitacion = 0
-				UPDATE LOS_TRIGGERS.Usuario set user_afiliado = null where user_afiliado is not null
+				update LOS_TRIGGERS.Afiliado set afil_habilitacion = 0
+				update LOS_TRIGGERS.Usuario set user_afiliado = null where user_afiliado is not null
 			END
 		ELSE IF (@nombre_rol='Administrador')
 			BEGIN
-				UPDATE LOS_TRIGGERS.Administrador set admi_habilitacion = 0
-				UPDATE LOS_TRIGGERS.Usuario set user_administrador = null where user_administrador is not null
+				update LOS_TRIGGERS.Administrador set admi_habilitacion = 0
+				update LOS_TRIGGERS.Usuario set user_administrador = null where user_administrador is not null
 			END
 		ELSE IF (@nombre_rol='Profesional')
 			BEGIN
-				UPDATE LOS_TRIGGERS.Profesional set prof_habilitacion = 0
-				UPDATE LOS_TRIGGERS.Usuario set user_profesional = null where user_profesional is not null
+				update LOS_TRIGGERS.Profesional set prof_habilitacion = 0
+				update LOS_TRIGGERS.Usuario set user_profesional = null where user_profesional is not null
 			END
 	END;
 GO
@@ -86,25 +127,17 @@ GO
 --- << Funcionalidades >> ---
 IF OBJECT_ID ('LOS_TRIGGERS.AgregarFuncionalidadAUnRol') is not null DROP PROCEDURE LOS_TRIGGERS.AgregarFuncionalidadAUnRol
 GO
-CREATE PROC LOS_TRIGGERS.AgregarFuncionalidadAUnRol (@funcionalidad varchar(255), @rol varchar(255)) AS
+CREATE PROC LOS_TRIGGERS.AgregarFuncionalidadAUnRol (@rol varchar(255), @funcionalidad varchar(255)) AS
 BEGIN
 		declare @func_id as numeric(18,0)
 		set @func_id = (select func_id from LOS_TRIGGERS.Funcionalidad where func_nombre=@funcionalidad)
+
 		IF (@rol='Afiliado')
-			BEGIN
-				insert into LOS_TRIGGERS.Funcionalidad_Rol
-					select @func_id, afil_numero from LOS_TRIGGERS.Afiliado
-			END
+			insert into LOS_TRIGGERS.Funcionalidad_Rol select @func_id, afil_numero from LOS_TRIGGERS.Afiliado
 		ELSE IF (@rol='Administrador')
-			BEGIN
-				insert into LOS_TRIGGERS.Funcionalidad_Rol
-					select @func_id, admi_id from LOS_TRIGGERS.Administrador
-			END
+			insert into LOS_TRIGGERS.Funcionalidad_Rol select @func_id, admi_id from LOS_TRIGGERS.Administrador
 		ELSE IF (@rol='Profesional')
-			BEGIN
-				insert into LOS_TRIGGERS.Funcionalidad_Rol
-					select @func_id, prof_id from LOS_TRIGGERS.Profesional
-			END
+			insert into LOS_TRIGGERS.Funcionalidad_Rol select @func_id, prof_id from LOS_TRIGGERS.Profesional
 END;
 GO
 
@@ -114,23 +147,38 @@ CREATE PROC LOS_TRIGGERS.QuitarFuncionalidadAUnRol (@funcionalidad varchar(255),
 BEGIN
 		declare @func_id as numeric(18,0)
 		set @func_id = (select func_id from LOS_TRIGGERS.Funcionalidad where func_nombre=@funcionalidad)
+
 		IF (@rol='Afiliado')
-			BEGIN
-				delete from LOS_TRIGGERS.Funcionalidad_Rol
-				where funcionalidad=@func_id AND rol IN (select afil_numero from LOS_TRIGGERS.Afiliado)
-			END
+			delete from LOS_TRIGGERS.Funcionalidad_Rol where funcionalidad=@func_id AND rol IN (select afil_numero from LOS_TRIGGERS.Afiliado)
 		ELSE IF (@rol='Administrador')
-			BEGIN
-				delete from LOS_TRIGGERS.Funcionalidad_Rol
-				where funcionalidad=@func_id AND rol IN (select admi_id from LOS_TRIGGERS.Administrador)
-			END
+			delete from LOS_TRIGGERS.Funcionalidad_Rol where funcionalidad=@func_id AND rol IN (select admi_id from LOS_TRIGGERS.Administrador)
 		ELSE IF (@rol='Profesional')
-			BEGIN
-				delete from LOS_TRIGGERS.Funcionalidad_Rol
-				where funcionalidad=@func_id AND rol IN (select prof_id from LOS_TRIGGERS.Profesional)
-			END
+			delete from LOS_TRIGGERS.Funcionalidad_Rol where funcionalidad=@func_id AND rol IN (select prof_id from LOS_TRIGGERS.Profesional)
 END;
 GO
+
+-- << Carga de Funcionalidades a los Roles >>
+EXEC LOS_TRIGGERS.AgregarFuncionalidadAUnRol @rol = 'Administrador', @funcionalidad = 'Login y Seguridad'
+EXEC LOS_TRIGGERS.AgregarFuncionalidadAUnRol @rol = 'Administrador', @funcionalidad = 'ABM de Rol'
+EXEC LOS_TRIGGERS.AgregarFuncionalidadAUnRol @rol = 'Administrador', @funcionalidad = 'ABM de Afiliado'
+EXEC LOS_TRIGGERS.AgregarFuncionalidadAUnRol @rol = 'Administrador', @funcionalidad = 'ABM de Profesional'
+EXEC LOS_TRIGGERS.AgregarFuncionalidadAUnRol @rol = 'Administrador', @funcionalidad = 'ABM de Plan'
+EXEC LOS_TRIGGERS.AgregarFuncionalidadAUnRol @rol = 'Administrador', @funcionalidad = 'ABM de Especialidad Médica'
+EXEC LOS_TRIGGERS.AgregarFuncionalidadAUnRol @rol = 'Administrador', @funcionalidad = 'Registro de Usuario'
+EXEC LOS_TRIGGERS.AgregarFuncionalidadAUnRol @rol = 'Administrador', @funcionalidad = 'Registro de Agenda de Profesional'
+EXEC LOS_TRIGGERS.AgregarFuncionalidadAUnRol @rol = 'Administrador', @funcionalidad = 'Compra de Bono'
+EXEC LOS_TRIGGERS.AgregarFuncionalidadAUnRol @rol = 'Administrador', @funcionalidad = 'Listado Estadístico'
+
+EXEC LOS_TRIGGERS.AgregarFuncionalidadAUnRol @rol = 'Profesional', @funcionalidad = 'Login y Seguridad'
+EXEC LOS_TRIGGERS.AgregarFuncionalidadAUnRol @rol = 'Profesional', @funcionalidad = 'Registro de Agenda de Profesional'
+EXEC LOS_TRIGGERS.AgregarFuncionalidadAUnRol @rol = 'Profesional', @funcionalidad = 'Registro de Consulta Médica'
+EXEC LOS_TRIGGERS.AgregarFuncionalidadAUnRol @rol = 'Profesional', @funcionalidad = 'Registro de Diagnóstico Médico'
+EXEC LOS_TRIGGERS.AgregarFuncionalidadAUnRol @rol = 'Profesional', @funcionalidad = 'Cancelación de Turno'
+
+EXEC LOS_TRIGGERS.AgregarFuncionalidadAUnRol @rol = 'Afiliado', @funcionalidad = 'Login y Seguridad'
+EXEC LOS_TRIGGERS.AgregarFuncionalidadAUnRol @rol = 'Afiliado', @funcionalidad = 'Compra de Bono'
+EXEC LOS_TRIGGERS.AgregarFuncionalidadAUnRol @rol = 'Afiliado', @funcionalidad = 'Pedido de Turno'
+EXEC LOS_TRIGGERS.AgregarFuncionalidadAUnRol @rol = 'Afiliado', @funcionalidad = 'Cancelación de Turno'
 
 --- << Pedido de un Turno >> ---
 IF OBJECT_ID ('LOS_TRIGGERS.ComboProfesionales') is not null DROP PROCEDURE LOS_TRIGGERS.ComboProfesionales
@@ -158,19 +206,19 @@ IF OBJECT_ID ('LOS_TRIGGERS.ComboHorariosDeUnProfesionalEnUnaEspecialidad') is n
 GO
 CREATE PROC LOS_TRIGGERS.ComboHorariosDeUnProfesionalEnUnaEspecialidad (@profesional numeric(18,0), @especialidad numeric(18,0)) AS
 	BEGIN
-		select hora_id, hora_fecha_asignada, hora_dia_atencion, hora_inicio_atencion
-		from LOS_TRIGGERS.Horario_Atencion
-		where hora_especialidad_profesional=(select espe_prof_id from LOS_TRIGGERS.Especialidad_Profesional where profesional=@profesional AND especialidad=@especialidad) AND hora_fecha_asignada >= GETDATE()
-		order by hora_fecha_asignada, hora_inicio_atencion
+		select dia_id, dia_nombre_dia, dia_hora_inicio, dia_hora_fin
+		from LOS_TRIGGERS.Dia_Atencion
+		where dia_especialidad_profesional=(select espe_prof_id from LOS_TRIGGERS.Especialidad_Profesional where profesional=@profesional AND especialidad=@especialidad)
+		order by dia_id
 	END;
 GO
 
 IF OBJECT_ID ('LOS_TRIGGERS.RegistrarTurno') is not null DROP PROCEDURE LOS_TRIGGERS.RegistrarTurno
 GO
-CREATE PROC LOS_TRIGGERS.CancelarTurno (@afiliado numeric(18,00), @profesional numeric(18,00), @fecha datetime) AS
+CREATE PROC LOS_TRIGGERS.CancelarTurno (@afiliado numeric(18,00), @profesional numeric(18,00), @especialidad numeric(18,0), @fecha datetime) AS
 BEGIN
-		insert into LOS_TRIGGERS.Turno (turn_afiliado, turn_profesional, turn_fecha_atencion)
-			values(@afiliado, @profesional, @fecha)
+		insert into LOS_TRIGGERS.Turno (turn_afiliado, turn_especialidad_profesional, turn_fecha)
+			select @afiliado, (select espe_prof_id from LOS_TRIGGERS.Especialidad_Profesional where profesional=@profesional AND especialidad=@especialidad), @fecha
 END
 GO
 
@@ -179,10 +227,10 @@ IF OBJECT_ID ('LOS_TRIGGERS.TurnosAsignadosAUnAfiliado') is not null DROP PROCED
 GO
 CREATE PROC LOS_TRIGGERS.TurnosAsignadosAUnAfiliado (@afiliado numeric(18,0)) AS
 	BEGIN
-		select turn_numero, user_apellido+', '+user_nombre, espe_descripcion, hora_fecha_asignada, hora_inicio_atencion
-		from LOS_TRIGGERS.Turno, LOS_TRIGGERS.Horario_Atencion, LOS_TRIGGERS.Usuario, LOS_TRIGGERS.Especialidad
-		where turn_afiliado=@afiliado AND hora_turno=turn_numero AND user_profesional=turn_profesional AND espe_codigo=turn_especialidad
-		order by hora_fecha_asignada, hora_inicio_atencion
+		select turn_numero, user_apellido+', '+user_nombre, espe_descripcion, CONVERT(date, turn_fecha), turn_nombre_dia, turn_hora_inicio
+		from LOS_TRIGGERS.Turno, LOS_TRIGGERS.Usuario, LOS_TRIGGERS.Especialidad_Profesional, LOS_TRIGGERS.Especialidad
+		where turn_afiliado=@afiliado AND turn_fecha >= GETDATE() AND espe_prof_id=turn_especialidad_profesional AND espe_codigo=especialidad
+		order by turn_fecha, turn_hora_inicio
 	END;
 GO
 
@@ -191,15 +239,14 @@ GO
 CREATE PROC LOS_TRIGGERS.CancelarTurno (@afiliado numeric(18,00), @profesional numeric(18,00), @turno numeric(18,0), @tipo numeric(18,0), @motivo varchar(255)) AS
 BEGIN
 	declare @fecha_turno as datetime
-	set @fecha_turno = (select turn_fecha_atencion from LOS_TRIGGERS.Turno where turn_numero=@turno)
+	set @fecha_turno = (select turn_fecha from LOS_TRIGGERS.Turno where turn_numero=@turno)
 
 	IF (GETDATE() < @fecha_turno)
 		BEGIN
-			insert into LOS_TRIGGERS.Cancelacion_Turno (canc_afiliado, canc_profesional, canc_fecha, canc_fecha_turno, canc_tipo, canc_motivo, can_horario_atencion)
-				select @afiliado, @profesional, @fecha_turno, GETDATE(), @tipo, @motivo, hora_id
-				from LOS_TRIGGERS.Horario_Atencion where hora_turno=@turno
+			insert into LOS_TRIGGERS.Cancelacion_Turno (canc_afiliado, canc_profesional, canc_fecha_y_hora, canc_fecha_turno, canc_tipo, canc_motivo)
+				select @afiliado, @profesional, @fecha_turno, GETDATE(), @tipo, @motivo
+	
 
-			update LOS_TRIGGERS.Horario_Atencion set hora_turno = null where hora_turno=@turno
 			delete from LOS_TRIGGERS.Turno where turn_numero=@turno
 		END
 END;
@@ -235,7 +282,7 @@ GO
 CREATE PROC LOS_TRIGGERS.RegistrarDiaAtencionProfesional (@profesional numeric(18,0), @especialidad numeric(18,0), @dia varchar(255), @inicio varchar(255), @fin varchar(255)) AS
 BEGIN
 	declare @horas_acumuladas as int
-	set @horas_acumuladas = (select prof_cant_horas_laborales from LOS_TRIGGERS.Profesional where prof_id=@profesional) + DATEDIFF(hour, @fin, @inicio)
+	set @horas_acumuladas = (select prof_horas_laborales from LOS_TRIGGERS.Profesional where prof_id=@profesional) + DATEDIFF(hour, @fin, @inicio)
 	
 	IF (@horas_acumuladas <= 48)
 		BEGIN
@@ -243,7 +290,7 @@ BEGIN
 				select @dia, @inicio, @fin, (select espe_prof_id from LOS_TRIGGERS.Especialidad_Profesional where profesional=@profesional AND especialidad=@especialidad)
 
 			update LOS_TRIGGERS.Profesional
-				set prof_cant_horas_laborales = CONVERT(numeric(2,0), @horas_acumuladas)
+				set prof_horas_laborales = CONVERT(numeric(2,0), @horas_acumuladas)
 		END
 	--ELSE
 		-- NO PERMITIDO
